Added xml exporting of an Ariadne tree.
This is still only the export without any options like without grants
etc etc.

However.. It's _fast_, don't load it into your webbrowser but make a small
template with: <a href="export.wddx.phtml">export</a> and do a save as of
that link.

A default (out of the box) Ariadne tree is exported and downloaded in 6 seconds.


git-svn-id: https://svn.muze.nl/svn/ariadne-php5/trunk@2505 7882d7ee-46e0-0310-a255-a28a61b628e4