add a specific template for language select on pshortcut. this template is called for disktemplates not ending in .phtml that are called on pshortcut without being available on disk for pshortcut. That is, when the shortcut does not have the language.
.phtml templates are not affected due to the loader setting
$ARCurrent->allnls = true for any template ending in .phtml. Other
templates on disk not defined on pshortcut caused this dialog to
spawn. This now forwards the call to the target for determination 
of a user.languageselect.html.

The case that specifically triggered the issue is dialog.edit.php
which has a $ARCurrent->allnls = true; at the top. But due to not
being defined on pshortcut and the shortcut not having the same
nls settings as the target the user.languageselect.html was spawned
in the backend of ariadne.

yay, long commit notice, but hey, reasons are everything for code!



git-svn-id: https://svn.muze.nl/svn/ariadne-php5/trunk@4298 7882d7ee-46e0-0310-a255-a28a61b628e4