Add SU style abilities to Ariadne.
pobject has been extended with 4 new functions 3 of which are exposed to pinp:

suKey( $grants ) - generates a sha1 hash for the given grantstring on the current object (not exposed to pinp, requires config grant)
suBegin( $grants, $key ) - matches the grantstring and they key on the path and adds the grants like groupgrants to the user for that path (and below)
suEnd() - removes all grants set by suBegin calls on the current path
suCall( $grants, $key, $function, $args) - convenience function wraps a regular call with suBegin and suEnd

Using this functionality requires you to set $AR->suSalt to a secret string in your ariadne.phtml config.

The ordering of the grantstring does not matter, it is parsed by mod_grant, sorted and serialized before the hash is generated.



git-svn-id: https://svn.muze.nl/svn/ariadne-php5/trunk@4319 7882d7ee-46e0-0310-a255-a28a61b628e4