Made list interface an abstract class.
Added optional modifier lists.

Added a 'share' modifier which you can use to claim (or share) a percentage of the
available count. (note that the engine always operates from right to left, so this
option should be used at the end of each (sub)expression).

Added a 'flow' modifier which you can use to change the flow (rl or lr) (lr is default).

Example:

  '(l .* n (c n .* r?)[share: 0.5])[flow: rl;]'


These changes are experimental: the language can change and features can get removed.



git-svn-id: https://svn.muze.nl/svn/ariadne-php5/trunk@4416 7882d7ee-46e0-0310-a255-a28a61b628e4